fix format and build problem v2

+++ tasks/mpi/rysev_m_gypercube/perf_tests/main.cpp
@@ -11,8 +11,8 @@ TEST(rysev_m_gypercube, test_pipeline_run) {
   boost::mpi::communicator world;
   std::vector<uint8_t> in(10000);
   std::vector<uint8_t> out(10000);
-  for (int i = 0; i < in.size(); i++) { 
-    in[i] = i % 256; 
+  for (size_t i = 0; i < in.size(); i++) {
+    in[i] = i % 256;
   }
 
   std::shared_ptr<ppc::core::TaskData> taskDataPar = std::make_shared<ppc::core::TaskData>();
@@ -21,7 +21,7 @@ TEST(rysev_m_gypercube, test_pipeline_run) {
   taskDataPar->outputs.emplace_back(reinterpret_cast<uint8_t *>(out.data()));
   taskDataPar->outputs_count.emplace_back(out.size());
 
-  auto testParTask = std::make_shared<rysev_m_gypercube::GyperCube>(taskDataPar);  
+  auto testParTask = std::make_shared<rysev_m_gypercube::GyperCube>(taskDataPar);
   ASSERT_TRUE(testParTask->validation());
   testParTask->pre_processing();
   testParTask->run();
@@ -35,8 +35,8 @@ TEST(rysev_m_gypercube, test_pipeline_run) {
   auto perfAnalyzer = std::make_shared<ppc::core::Perf>(testParTask);
   perfAnalyzer->pipeline_run(perfAttr, perfResults);
 
-  for (size_t i = 0; i < out.size(); i++) { 
-    ASSERT_EQ(out[i], in[i]); 
+  for (size_t i = 0; i < out.size(); i++) {
+    ASSERT_EQ(out[i], in[i]);
   }
   ppc::core::Perf::print_perf_statistic(perfResults);
 }
@@ -45,8 +45,8 @@ TEST(rysev_m_gypercube, test_task_run) {
   boost::mpi::communicator world;
   std::vector<uint8_t> in(10000);
   std::vector<uint8_t> out(10000);
-  for (int i = 0; i < in.size(); i++) { 
-    in[i] = i % 256; 
+  for (size_t i = 0; i < in.size(); i++) {
+    in[i] = i % 256;
   }
 
   std::shared_ptr<ppc::core::TaskData> taskDataPar = std::make_shared<ppc::core::TaskData>();
@@ -55,7 +55,7 @@ TEST(rysev_m_gypercube, test_task_run) {
   taskDataPar->outputs.emplace_back(reinterpret_cast<uint8_t *>(out.data()));
   taskDataPar->outputs_count.emplace_back(out.size());
 
-  auto testParTask = std::make_shared<rysev_m_gypercube::GyperCube>(taskDataPar);  
+  auto testParTask = std::make_shared<rysev_m_gypercube::GyperCube>(taskDataPar);
   ASSERT_TRUE(testParTask->validation());
   testParTask->pre_processing();
   testParTask->run();
@@ -69,8 +69,8 @@ TEST(rysev_m_gypercube, test_task_run) {
   auto perfAnalyzer = std::make_shared<ppc::core::Perf>(testParTask);
   perfAnalyzer->task_run(perfAttr, perfResults);
 
-  for (size_t i = 0; i < out.size(); i++) { 
-    ASSERT_EQ(out[i], in[i]); 
+  for (size_t i = 0; i < out.size(); i++) {
+    ASSERT_EQ(out[i], in[i]);
   }
   ppc::core::Perf::print_perf_statistic(perfResults);
 }
